requirements.txt
contourpy==1.3.0
cycler==0.12.1
filelock==3.13.1
fonttools==4.55.0
fsspec==2024.2.0
imageio==2.36.1
importlib_resources==6.4.5
Jinja2==3.1.3
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.9.3
monai==1.4.0
monai-generative==0.2.3
mpmath==1.3.0
networkx==3.2.1
nibabel==5.3.2
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
packaging==24.2
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
scikit-image==0.24.0
scipy==1.13.1
six==1.16.0
sympy==1.13.1
tifffile==2024.8.30
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1+cu118
tqdm==4.67.1
triton==3.1.0
typing_extensions==4.9.0
zipp==3.21.0
